--- backend/src/main/java/com/example/backend/identity/security/config/handler/CustomAuthenticationFailureHandler.java
@@ -1,31 +1,0 @@
-package com.example.backend.identity.security.config.handler;
-
-import java.io.IOException;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@Component
-public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
-	private final HandlerExceptionResolver exceptionResolver;
-
-	CustomAuthenticationFailureHandler(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver exceptionResolver) {
-		this.exceptionResolver = exceptionResolver;
-	}
-
-	@Override
-	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
-		AuthenticationException authException) throws IOException, ServletException {
-		// 내부적인 로그를 찍어보려면 이곳에서 하면 됨!
-
-		// AuthenticationException은 401에러로 GlobalExceptionHandler에서 한번에 처리
-		exceptionResolver.resolveException(request, response, null, authException);
-	}
-}
